requirements.txt
python-telegram-bot[webhooks]==22.6
Flask==3.1.0
gunicorn==23.0.0